conflict 해결

#141

+++ src/main/java/com/example/backoffice/domain/vacation/facade/VacationsServiceFacadeImplV1.java
@@ -7,7 +7,6 @@ import com.example.backoffice.domain.member.service.MembersServiceV1;
 import com.example.backoffice.domain.notification.converter.NotificationsConverter;
 import com.example.backoffice.domain.notification.entity.NotificationType;
 import com.example.backoffice.domain.notification.facade.NotificationsServiceFacadeV1;
-import com.example.backoffice.domain.notification.service.NotificationsServiceV1;
 import com.example.backoffice.domain.vacation.converter.VacationsConverter;
 import com.example.backoffice.domain.vacation.dto.VacationDateRangeDto;
 import com.example.backoffice.domain.vacation.dto.VacationsRequestDto;
@@ -31,14 +30,14 @@ import java.util.List;
 @RequiredArgsConstructor
 public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
     private final MembersServiceV1 membersService;
-    private final NotificationsServiceV1 notificationsService;
+    private final NotificationsServiceFacadeV1 notificationsServiceFacade;
     private final VacationsServiceV1 vacationsService;
     private final VacationPeriodHolder vacationPeriodHolder;
 
     // 휴가 관련 피드백 : 생성할 때, 승인이 될 시에, 잔여 휴가 일 변경
     // 날짜가 시작 날이 토요일/ 일요일은 안됨, 마지막일 또한 토요일/ 일요일은 안됨.
     @Override
-    public VacationsResponseDto.UpdatePeriodDto updatePeriod(
+    public VacationsResponseDto.UpdatePeriodDto updatePeriodByAdmin(
             Members loginMember, VacationsRequestDto.UpdatePeriodDto requestDto){
         validateUpdatePermission(loginMember);
 
@@ -65,7 +64,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         List<Members> memberList = membersService.findAll();
 
         for(Members member : memberList){
-            notificationsService.generateEntityAndSendMessage(
+            notificationsServiceFacade.createOne(
                     NotificationsConverter.toNotificationData(
                             member, loginMember, null, null, null, null,
                             "변경된 휴가 신청 기간 안내입니다 : "
@@ -79,7 +78,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
 
     @Override
     @Transactional
-    public VacationsResponseDto.CreateOneDto createOne(
+    public VacationsResponseDto.CreateOneDto createOneByMember(
             Members loginMember, VacationsRequestDto.CreateOneDto requestDto) {
 
         // 1. 기간 범위 검증
@@ -98,10 +97,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         vacationsService.save(vacation);
 
         if(requestDto.getUrgent()){
-            sendUrgentOneForHRManager(loginMember);
-            if(vacationType.equals(VacationType.ANNUAL_LEAVE)){
-                throw new VacationsCustomException(VacationsExceptionCode.DO_NOT_NEED_URGENT);
-            }
+            sendUrgentOneByHrManager(loginMember);
         }
 
         return VacationsConverter.toCreateOneDto(vacation);
@@ -109,7 +105,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
 
     @Override
     @Transactional(readOnly = true)
-    public VacationsResponseDto.ReadDayDto readDay(Long vacationId, Members loginMember) {
+    public VacationsResponseDto.ReadDayDto readDayByMember(Long vacationId, Members loginMember) {
         Vacations vacation = vacationsService.findById(vacationId);
         if (vacation.getOnVacationMember().getId().equals(loginMember.getId())) {
             throw new VacationsCustomException(VacationsExceptionCode.NO_PERMISSION_TO_READ_VACATION);
@@ -119,10 +115,14 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
 
     @Override
     @Transactional
-    public VacationsResponseDto.UpdateOneDto updateOne(
+    public VacationsResponseDto.UpdateOneDto updateOneByMember(
             Long vacationId, Members loginMember, VacationsRequestDto.UpdateOneDto requestDto) {
 
         Vacations vacation = vacationsService.findById(vacationId);
+
+        if(vacation.getIsAccepted()){
+            throw new VacationsCustomException(VacationsExceptionCode.NO_PERMISSION_TO_UPDATE_VACATION);
+        }
 
         VacationType vacationType = VacationsConverter.toVacationType(requestDto.getVacationType());
         validateVacationType(vacationType, requestDto.getUrgent());
@@ -137,7 +137,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
                 requestDto.getEndDate(), requestDto.getUrgent());
 
         if(requestDto.getUrgent()){
-            sendUrgentOneForHRManager(loginMember);
+            sendUrgentOneByHrManager(loginMember);
             if(vacationType.equals(VacationType.ANNUAL_LEAVE)){
                 throw new VacationsCustomException(VacationsExceptionCode.DO_NOT_NEED_URGENT);
             }
@@ -151,10 +151,9 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         return VacationsConverter.toUpdateOneDto(vacation);
     }
 
-    // 승인 요청만 받는 형태이기에 Admin의 RequestDto를 받을 필요가 없다.
     @Override
     @Transactional
-    public VacationsResponseDto.UpdateOneForAdminDto updateOneForAdmin(
+    public VacationsResponseDto.UpdateOneByAdminDto updateOneByAdmin(
             Long vacationId, Members loginMember){
         // 1. 로그인 멤버가 CEO || HR_MANAGER인지 확인
         validateUpdatePermission(loginMember);
@@ -171,7 +170,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
                 ? vacation.getOnVacationMember().getMemberName() + "님의 휴가가 승인되었습니다."
                 : vacation.getOnVacationMember().getMemberName() + "님의 휴가가 미승인되었습니다.";
 
-        notificationsService.generateEntityAndSendMessage(
+        notificationsServiceFacade.createOne(
                 NotificationsConverter.toNotificationData(
                         loginMember, vacation.getOnVacationMember(),
                         null, null, null, null,
@@ -201,30 +200,32 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
             membersService.minusVacationDays(vacation.getOnVacationMember(), vacationDays);
         } else {
             // 미승인된 경우, 휴가 일 수 복원 (즉, 다시 추가)
-            // 휴가 db에서 삭제
             membersService.addVacationDays(vacation.getOnVacationMember(), vacationDays);
-            vacationsService.deleteById(vacationId);
         }
 
         // 6. DTO 전송
-        return VacationsConverter.toUpdateOneForAdminDto(vacation);
+        return VacationsConverter.toUpdateOneByAdminDto(vacation);
     }
 
     @Override
     @Transactional
-    public void deleteOne(Long vacationId, Members loginMember) {
+    public void deleteOneByMember(Long vacationId, Members loginMember) {
         Vacations vacation = vacationsService.findById(vacationId);
         validateMemberPermission(
                 loginMember, vacation.getOnVacationMember().getId(),
                 null, VacationCrudType.DELETE_VACATION);
 
+        // 로그인 사용자가 해당 휴가라는 사실을 위에서 점검
+        if((vacation.getIsAccepted() == Boolean.TRUE)){
+            throw new VacationsCustomException(VacationsExceptionCode.NOT_ACCEPTED_DELETE_VACATION);
+        }
         vacationsService.deleteById(vacationId);
     }
 
     // 필터링된 휴가 관리 시스템 입장 시, 보이는 휴가 리스트
     @Override
     @Transactional(readOnly = true)
-    public List<VacationsResponseDto.ReadMonthDto> readForHrManager(
+    public List<VacationsResponseDto.ReadMonthDto> readByHrManager(
             Long year, Long month, Boolean isAccepted, Boolean urgent,
             String department, Members loginMember){
 
@@ -232,8 +233,8 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         membersService.findHRManagerOrCEO(loginMember);
 
         // 2. 시작일과 마감일을 DateTimeUtils에서 가져오기
-        LocalDateTime startDate = DateTimeUtils.getStartDayOfMonth(year, month);
-        LocalDateTime endDate = DateTimeUtils.getEndDayOfMonth(year, month); // 다음 달 1일 바로 전
+        LocalDateTime startDayOfMonth = DateTimeUtils.getStartDayOfMonth(year, month);
+        LocalDateTime endDayOfMonth = DateTimeUtils.getEndDayOfMonth(year, month); // 다음 달 1일 바로 전
 
         // 3. 초기 상태와 필터링 상태에 따른 리턴 형태 변경
         MemberDepartment memberDepartment = null;
@@ -241,8 +242,8 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         // 3-1. 초기 상태 리턴
         if (department == null && urgent == null && isAccepted == null){
             List<Vacations> vacationList
-                    = vacationsService.findVacationsOnMonth(startDate, endDate);
-            return VacationsConverter.toReadMonthForHrManager(vacationList);
+                    = vacationsService.findVacationsOnMonth(startDayOfMonth, endDayOfMonth);
+            return VacationsConverter.toReadMonthByHrManager(vacationList);
         }
 
         // 3-2 department가 null이 아닐 때 필터링 상태 리턴
@@ -253,19 +254,19 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         // 3-3 department가 null이면 모든 부서의 값을 가져온 상태로 리턴
         List<Vacations> vacationList
                 = vacationsService.findFilteredVacationsOnMonth(
-                        startDate, endDate, isAccepted, urgent, memberDepartment);
+                startDayOfMonth, endDayOfMonth, isAccepted, urgent, memberDepartment);
 
-        return VacationsConverter.toReadMonthForHrManager(vacationList);
+        return VacationsConverter.toReadMonthByHrManager(vacationList);
     }
 
     @Override
     @Transactional
-    public void deleteOneForHrManager(
-            Long vacationId, VacationsRequestDto.DeleteOneForAdminDto requestDto,
+    public void deleteOneByHrManager(
+            Long vacationId, VacationsRequestDto.DeleteOneByAdminDto requestDto,
             Members loginMember){
         membersService.findHRManagerOrCEO(loginMember);
         Vacations vacation = vacationsService.findById(vacationId);
-        notificationsService.generateEntityAndSendMessage(
+        notificationsServiceFacade.createOne(
                 NotificationsConverter.toNotificationData(
                         loginMember, vacation.getOnVacationMember(),
                         null, null, null, null,
@@ -346,8 +347,8 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         // 1. 모든 멤버의 수에 비례에서 휴가 사용
         if (memberTotalCount > 10) {
             for (long i = 0; i < vacationDays; i++) {
-                LocalDateTime customStartDate = startDate.plusDays(i);
-                long vacationingMembersCount = vacationsService.countVacationingMembers(customStartDate);
+                LocalDateTime currentDate = startDate.plusDays(i);
+                long vacationingMembersCount = vacationsService.countVacationingMembers(currentDate);
 
                 double vacationRate = (double) vacationingMembersCount / memberTotalCount;
 
@@ -378,12 +379,11 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
         }
     }
 
-    private void sendUrgentOneForHRManager(Members loginMember) {
-        Members admin = membersService.findHRManagerOrCEO(loginMember);
-
-        notificationsService.generateEntityAndSendMessage(
+    private void sendUrgentOneByHrManager(Members loginMember) {
+        Members hrManager = membersService.findHRManager();
+        notificationsServiceFacade.createOne(
                 NotificationsConverter.toNotificationData(
-                        admin, loginMember, null, null, null, null, null),
+                        hrManager, loginMember, null, null, null, null, null),
                 NotificationType.URGENT_VACATION);
     }
 
@@ -406,9 +406,7 @@ public class VacationsServiceFacadeImplV1 implements VacationsServiceFacadeV1{
                 break;
             }
             case DELETE_VACATION: {
-                if (!loginMember.getId().equals(onVacationMemberId)
-                        && !(loginMember.getPosition().equals(MemberPosition.MANAGER)
-                        && loginMember.getDepartment().equals(MemberDepartment.HR))) {
+                if (!loginMember.getId().equals(onVacationMemberId)) {
                     throw new VacationsCustomException(VacationsExceptionCode.NO_PERMISSION_TO_DELETE_VACATION);
                 }
                 break;
